split up macro and factory defaults registration to plone 4 and 5
