+++ src/yafowil/plone/profiles/plone5/registry.xml
@@ -10,4 +10,12 @@
     <value key="depends">plone</value>
   </records>
 
+  <!-- bootstrap glyphicons -->
+  <record name="bootstrap.glyphicons">
+    <field type="plone.registry.field.Bool">
+      <title>Glyphicons Halflings Web Font</title>
+    </field>
+    <value>True</value>
+  </record>
+
 </registry>
